Split BillingServiceTests along the billingStateService boundary
BillingServiceTests.swift had grown to 1023 lines with both
type_body_length and file_length disabled outright. Split it into:
- BillingServiceTests.swift: checkout, plan/subscription lookups,
  self-hosted detection, and premiumStatusChanged — none of which touch
  billingStateService.
- BillingService+BillingStateTests.swift: action cards, the
  subscription attention card, and premium-upgrade-pending
  reconciliation — everything that does, taking the
  PremiumUpgradeStateStore/resolvedUserId mock-wiring helpers with it
  since they're now only used here.

Follows the existing TypeName+ConcernTests.swift convention (separate
type, duplicated setup) already used for VaultListProcessor+BillingTests.swift.
